requirements.txt
google-generativeai
google-cloud-bigquery
google-auth
db_dtypes